+++ Changes.md
@@ -1,3 +1,7 @@
+# v1.0.0-beta2
+- Add a suppressNotSupportedException parameter for creating IProgressNotification.
+- Add INotificationManager.Show() to show a notification quickly.
+
 # v1.0.0-beta1
 - A C# library for sending native system notifications across desktop platforms.
 - Supports Windows (toast notifications), Linux (freedesktop.org), and macOS (NSUserNotification).